Remove archiving workaround for PR edits
Prior to this commit we would archive the existing card if we wanted to
edit the PR fields due to a bug in the trello ruby library.
This commit removes that workaround because the issue in the library has
been fixed. Additionally, it changes `create_trello_card` to fetch the
existing card.